requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# Paper parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
markdown==3.5.1

# LLM integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
langchain==0.1.0
tiktoken==0.5.2

# Data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Testing and validation
pytest==7.4.3
hypothesis==6.92.1

# Code execution
jupyter==1.0.0
ipykernel==6.27.1

# Output generation
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0
